+++ app/controllers/idv/ssn_controller.rb
@@ -2,6 +2,7 @@ module Idv
   class SsnController < ApplicationController
     include IdvSession
     include IdvStepConcern
+    include OutageConcern
     include StepIndicatorConcern
     include StepUtilitiesConcern
     include Steps::ThreatMetrixStepHelper
@@ -11,6 +12,7 @@ module Idv
     before_action :confirm_document_capture_complete
     before_action :confirm_repeat_ssn, only: :show
     before_action :override_csp_for_threat_metrix_no_fsm
+    before_action :check_for_outage, only: :show
 
     attr_accessor :error_message
