+++ app/controllers/idv/verify_info_controller.rb
@@ -1,6 +1,7 @@
 module Idv
   class VerifyInfoController < ApplicationController
     include IdvStepConcern
+    include OutageConcern
     include StepUtilitiesConcern
     include StepIndicatorConcern
     include VerifyInfoConcern
@@ -8,6 +9,7 @@ module Idv
 
     before_action :confirm_ssn_step_complete
     before_action :confirm_verify_info_step_needed
+    before_action :check_for_outage, only: :show
 
     def show
       @step_indicator_steps = step_indicator_steps
